Refactor: StackedBarChart의 TableView 분리
